Compiler warning fixes (#72)
-using new compiler flags

+++ shell/main.cc
@@ -84,7 +84,7 @@ int main(int argc, char** argv) {
 
   FML_LOG(INFO) << GIT_BRANCH " @ " GIT_HASH;
 
-  config.view.vm_args.reserve(argc - 1);
+  config.view.vm_args.reserve(static_cast<unsigned long>(argc - 1));
   for (int i = 1; i < argc; ++i) {
     config.view.vm_args.emplace_back(argv[i]);
   }
